Hi,

This improves the speed of subset by following a recommendation in the TODO comments and other recursion improvements.

These are the results

![image](https://github.com/user-attachments/assets/871182c1-8a1f-4c9f-9ad5-a22dd4e0ab5b)

```
develop
======
subset(matrix, rows and columns)         1.89 µs   ±0.13%
subset(array, rows and columns)          4.89 µs   ±0.25%
subset(matrix, booleans and columns)     2.15 µs   ±0.25%
subset(array, booleans and columns)      5.09 µs   ±0.14%
subset(matrix, columns and booleans)     2.15 µs   ±0.35%
subset(array, columns and booleans)      5.13 µs   ±0.34%
subset(matrix, range and rows)           1.19 µs   ±0.20%
subset(array, range and rows)            4.19 µs   ±0.16%
subset(matrix, range and all)            1.73 µs   ±0.20%
subset(array, range and all)             4.70 µs   ±0.16%
subset(matrix, all and range)            2.32 µs   ±0.20%
subset(array, all and range)             5.24 µs   ±0.19%
subset(matrix, all and all)              4.69 µs   ±0.26%
subset(array, all and all)               7.64 µs   ±0.20%

this pr
=====
subset(matrix, rows and columns)         1.23 µs   ±0.15%
subset(array, rows and columns)          4.17 µs   ±0.19%
subset(matrix, booleans and columns)     1.37 µs   ±0.18%
subset(array, booleans and columns)      4.30 µs   ±0.50%
subset(matrix, columns and booleans)     1.37 µs   ±0.18%
subset(array, columns and booleans)      4.29 µs   ±0.20%
subset(matrix, range and rows)           0.82 µs   ±0.20%
subset(array, range and rows)            3.74 µs   ±0.16%
subset(matrix, range and all)            0.81 µs   ±0.26%
subset(array, range and all)             3.64 µs   ±0.14%
subset(matrix, all and range)            1.10 µs   ±0.37%
subset(array, all and range)             3.97 µs   ±0.30%
subset(matrix, all and all)              1.64 µs   ±0.20%
subset(array, all and all)               4.43 µs   ±0.17%
```