@josdejong 2025-05-07
@dvd101x so what is the current status? Does the PR still yield good performance improvements after your latest commits, returning to use `x.map`?