@dvd101x 2025-05-15
Hi Jos, thanks for the response. I will just add a big matrix benchmark and I think that will be it.

After your comments and the benchmarks, I think the current implementations are ok. Returning the use of map will benefit readability and would affect performance for big matrices in chrome and in safari a bit less. Firefox would be still ok.

The current logic after many tests I think it's ok to follow this suggestion:
- for stuff that could be big (like matrices or arrays) avoid the use of `array.map` and `array.forEach`, and implement something like a for as done here or on `matrix.map` and `matrix.forEach`.
- For small stuff like the size of an array or an index I think it's ok to use `array.map` and `array.forEach`

So I will only add a benchmark for a bigMatrix and all the comments will be addressed.

Thanks for your patience and follow up.