@dvd101x 2025-05-15
Hi, I included a few benchmarks for bigger matrices, and the difference is also significant. More so when the ranges for the subset are further away from 0. In which case the benefits are some of the highest.

![image](https://github.com/user-attachments/assets/c7229490-244e-47cc-a150-5e38ef213c69)

With this I think all comments are addressed and it's ready to be merged.

```
PR 3467

subset(bigMatrix, rows and columns)                   1.42 µs   ±0.34%
subset(bigArray, rows and columns)                    1.40 µs   ±0.30%
subset(bigMatrix, range and columns)                  1.74 µs   ±0.88%
subset(bigArray, rows and range)                      1.84 µs   ±0.38%

develop

subset(bigMatrix, rows and columns)                   1.93 µs   ±0.30%
subset(bigArray, rows and columns)                    1.95 µs   ±1.64%
subset(bigMatrix, range and columns)                  3.29 µs   ±0.47%
subset(bigArray, rows and range)                      4.06 µs   ±0.24%
```