@josdejong 2025-05-21
👌 Nice, I like how this turned out, this is a significant improvement indeed! Thanks David